export-import düzeni değişti. Metotlarda düzenlemeler yapıldı. postman collection eklendi.

+++ 1-start-todos/repositories/boardsRepository.js
@@ -1,16 +1,32 @@
 import Board from '../models/boardsModel.js';
 import UserBoard from '../models/userBoardsModel.js';
 
-const createBoard = async (boardName) => {
-    const newBoard = new Board({ boardName });
+export const createBoard = async (userId, name) => {
+    const newBoard = new Board();
+    newBoard.adminUserId = userId;
+    newBoard.name = name;
     await newBoard.save();
+
+    const newUserBoard = new UserBoard();
+    newUserBoard.userId = userId;
+    newUserBoard.boardId = newBoard._id;
+    await newUserBoard.save();
     return newBoard;
 };
 
-const deleteBoard = async (boardId) => {
+export const deleteBoard = async (userId, boardId) => {
+    const board = await Board.findById(boardId);
+
+    if (!board) {
+        throw new Error('Board not found.');
+    }
+
+    if (board.adminUserId.toString() !== userId.toString()) {
+        throw new Error('Only the admin user can delete this board.');
+    }
+
     const deletedBoard = await Board.findOneAndDelete({ _id: boardId });
 
-    // Silinen board için userBoard tablosundan ilgili kayıtları temizle
     if (deletedBoard) {
         await UserBoard.deleteMany({ boardId: boardId });
     }
@@ -18,19 +34,34 @@ const deleteBoard = async (boardId) => {
     return deletedBoard;
 };
 
-const updateBoard = async (boardId, name) => {
+export const updateBoard = async (userId, boardId, name) => {
+    const board = await Board.findById(boardId);
+
+    if (!board) {
+        throw new Error('Board not found.');
+    }
+
+    if (board.adminUserId.toString() !== userId.toString()) {
+        throw new Error('Only the admin user can update this board.');
+    }
+
     const updatedBoard = await Board.findOneAndUpdate({ _id: boardId }, { name }, { new: true });
     return updatedBoard;
 };
 
-const getAllBoards = async (userId) => {
+export const getAllBoards = async (userId) => {
     const userBoards = (await UserBoard.find({ userId: userId }));
+
+    if (!userBoards) {
+        throw new Error('User\'s boards not found.');
+    }
+
     const boardIds = userBoards.map(ub => ub.boardId);
     const boards = await Board.find({ _id: { $in: boardIds } }).populate('todos');
     return boards;
 };
 
-const getBoardById = async (userId, boardId) => {
+export const getBoardById = async (userId, boardId) => {
     const userBoard = await UserBoard.findOne({ userId: userId, boardId: boardId });
 
     if (!userBoard) {
@@ -41,7 +72,7 @@ const getBoardById = async (userId, boardId) => {
     return board;
 };
 
-const getTodoById = async (userId, boardId, todoId) => {
+export const getTodoById = async (userId, boardId, todoId) => {
     const userBoard = await UserBoard.findOne({ userId: userId, boardId: boardId });
 
     if (!userBoard) {
@@ -50,13 +81,4 @@ const getTodoById = async (userId, boardId, todoId) => {
 
     const todo = await Board.findOne({ _id: boardId, todos: todoId }).populate('todos');
     return todo;
-};
-
-export default {
-    createBoard,
-    deleteBoard,
-    updateBoard,
-    getAllBoards,
-    getBoardById,
-    getTodoById
 };
